AGENT/requirements.txt

3️⃣ **requirements.txt** (Python dependencies):
```plaintext:D%3A%5CReact%20learning%5CHOSPITAL%5Chospital-ai-receptionist%5Crequirements.txt
Flask==3.0.2
python-dateutil==2.8.2
spacy==3.7.4
python-dotenv==1.0.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl